requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlmodel==0.0.22
apscheduler==3.10.4
playwright==1.49.0
python-dotenv==1.0.1
pillow==11.0.0
openai==1.57.0
anthropic==0.40.0
httpx==0.28.0
